Updated dependencies and deploy process
This is a bit of an overhaul of the library with the main aim being to
include eslint/prettier as dependencies so we don't have to directly
include them in other projects and deal with updates there.

See https://github.com/fishbrain/eslint-config-fishbrain-base/pull/469
for more details.

I also updated the CI scripts while I was here to match our flows in other repos.